requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
sqlalchemy==2.0.35
python-dotenv==1.0.1
bcrypt==4.2.0
python-jose[cryptography]==3.3.0
pydantic==2.9.2
python-multipart==0.0.9
psycopg2-binary==2.9.11
